Rename VError to BaseForm
